pyproject.toml
[tool.poetry]
name = "MosObserver"
version = "0.1.0"
description = ""
authors = ["rsaleev <rs.aleev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.1"
python-dotenv = "^0.20.0"
tortoise-orm = "^0.19.1"
aiohttp = "^3.8.1"
uvloop = "^0.16.0"
aiogram = "^2.20"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
